Updated SelectTable.java
Changed line 39 so that it would return double instead of integer. Fixes formatting problem when database data from price would display incorrectly.

+++ dbrun/src/dbrun/SelectTable.java
@@ -36,7 +36,7 @@ public class SelectTable {
             while (rs.next()) {  
                 System.out.println(rs.getInt("ItemNum") +  "\t" +   
                                    rs.getString("ItemName") + "\t" +  
-                                   rs.getInt("Price") + "\t" +
+                                   rs.getDouble("Price") + "\t" +
                                    rs.getString("Size")
                );  
             }
